requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
asyncpg
fastapi-users[sqlalchemy]
aiomqtt
python-multipart
python-dotenv
paho-mqtt